refactor: room 관련 컨트롤러 파일 이동 (app.controller -> room.controller)
